requirements.txt
numpy==1.9.2
pandas==0.16.0
progressbar2==2.7.3
PyYAML==3.11
requests==2.6.0
scikit-learn==0.16.0
xgboost==0.32
-e git+https://github.com/ldamewood/pystats.git#egg=pystats
